pyproject.toml
[project]
name = "verbs-play-bot"
version = "0.1.0"
description = "Verbs Play Customer Service Bot"
requires-python = ">=3.13"
dependencies = [
    "editorconfig-checker>=3.4.0",
    "google-api-core>=2.26.0",
    "google-cloud-dialogflow>=2.41.2",
    "grpcio>=1.75.1",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "ruff>=0.14.0",
]
